Start using doc-gen4 for API-level documentation (#7)
*Description of changes:*

Doc-gen4 is now used to generate API-level documentation in `docs/api`
to encourage the addition and formatting of documentation strings in
Strata source code.

NOTE: Docs. are not generated as a part of the CI yet.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

Co-authored-by: Shilpi Goel <shilgoel@amazon.com>