Change default names for assertions to avoid crashes with z3.  (#10)
*Description of changes:*

The default name for assertions was a string version of the assertion
itself. However, z3 crashes when the `smt2` filename contains `=`.
We create a `smt2_filename` function to centralize the logic. This
function returns an `.smt2` filename, replacing `=`, `==`, and ` ` by
`-`, `eq`, `_` respectively.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.