Add boolean literals to `ArithPrograms` (#22)
*Description of changes:*

Previously, `ArithPrograms` overloaded non-zero numbers as `true` and
zero as `false`, which didn't map easily to SMTLIB. This PR introduces
boolean literals to `ArithPrograms`, which allows the evaluator to
return `true` or `false` for equalities.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

Co-authored-by: Shilpi Goel <shilgoel@amazon.com>