Simpler and more complete output from StrataVerify (#20)
This changes `StrataVerify` to print only a few messages, including the
status of each goal and the overall summary of proof success. One
implication is that we don't discard goals that were discharged by
evaluation (because we still want to communicate that they're valid).

Right now, this means trivial goals are being sent to the solver, though
it would be easy to check for them in advance to remove that small
overhead.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.