Add axiomatization infrastructure for Boogie functions, add axioms for maps functions (#17)
*Description of changes:*

`update` and `select` on Maps in the Boogie languages are defined by
axioms. To add them, this PR adds a mechanism to add axioms for
functions in the Boogie functions factory. These axioms are encoded in
SMT for each type instantiations when encoding the functions.


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.