Refactor the Lambda and Boogie Factory contents (#36)
This reduces duplication and adds more explicit type annotations.

It's part of #35, extracted because it's not related to loops and is
independently useful. The changes to the files included in #35 are
identical except that it doesn't add loop translation to
`Strata.Languages.Boogie.DDMTransform.Translate`.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.