Rename StrataDD to StrataDDL and minor cleanups. (#40)
This renames the StrataDD dialect to StrataDDL to reflect it is for the
dialect definition language.

It also make a few minor cleanups to architecture document and info
tree.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.