Call Elimination Correctness Proof (#37)
*Description of changes:*

This PR adds the semantic-preserving proof for the call-elimination
transformation of Boogie programs.

Additionally, it:
- adds both call-elimination and deterministic-to-nondeterministic
correctness proofs to Strata root file to be checked upon compilation.
- fixes errors in nondeterministic transformation proof introduced
by the `loop` construct, and fixes the admits in `SemanticsProps.lean`.

The following are admitted and left as future work:
- Proofs that certain well-formedness conditions are checked by the
type-checker
- Proofs that the String generator and Boogie Identifier generator
produce globally unique labels
- Proofs that the Boogie evaluation semantics (`EvalStatement`) can be
approximated by the contract semantics (`EvalStatementContract`).

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: David Deng <htd@amazon.com>
Co-authored-by: Shilpi Goel <shigoel@gmail.com>