Rename Strata.Environment to Strata.Program and adjust formatting (#44)
This renames Strata.Environment to Strata.Program.

It also adjusts serialization and formatting for consistency and
consistent content-based disambiguation of dialects and programs.

This affects many files as it changes the pretty-printed representation
of programs.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.