Support for type alias in LExprT (temporary fix, if needed) (#39)
*Description of changes:*

This adds supports for Type aliases in type annotations on LExpr.
Ideally, we'd like to parametrized LExpr on the type of type
annotations, to let users write LExpr with LTy, and have LExpr with only
LMonoTy later in the type analysis and partial evaluation.
This PR fixes the bug in case we need a temporary fix.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: Andrew Wells <130512013+andrewmwells-amazon@users.noreply.github.com>