Refactoring: LExpr parametrized by the type of type (#52)
*Description of changes:*

Parametrized `LExpr` on the type of type they are carrying (i.e.
`LMonoTy` and `LTy`). This is a first step. The end goal is to have
users write `LExpr LTy` only, and then have a phase that translates
`LExpr LTy` into `LExpr LMonoTy` that will then be type checked.

This started by an observation that type checking is extremely slow and
that it seems to be (at least partly) due to the de-aliasing of types
that is happening often during current type checking, resulting in a
vast majority of the calls to `Contraints.unify` slowing the process
down considerably.

Next steps
- [ ] Write the transformation phase `LExpr LTy` -> `LExpr LMonoTy`
- [ ] Boogie Strata to generate `LExpr LTy`, and invoke the
transformation phase

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.