Fix extremely inefficient hash function leading to exponential SMT Encoding (#68)
*Description of changes:*

@MikaelMayer and I found the source of the slowness of the SMT Encoding:

The previous hash function on `Term` was computing the string
representation of the term before hashing it. As this is done for all
terms in `encodeTerm` to check there presence in the environment hash
tables, this introduced an exponential complexity for the SMT Encoder.

With this new hash function which is extremely efficient, the time to
verify one of my benchmark goes down from 38sec to 1.2sec.


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: Andrew Wells <130512013+andrewmwells-amazon@users.noreply.github.com>