Add back a commented out test (#63)
*Description of changes:*

This test was commented out because of missing type aliases support, but
with PR #39, it works.

Also, I added a label to the assertion, otherwise the `.smt2` filename
is too long and is rejected by `z3`. I will file an issue.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.