Rename DeclBinding types to ArgDecl. (#62)
The previous name was introduced when these were used for generic
bindings and not just op and function argument declarations.

The new name reflects what these are - declarations of arguments to
operators and functions.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

Co-authored-by: Shilpi Goel <shigoel@gmail.com>