Heap encoding in Boogie.Strata POC (#34)
Add an example of Strata.Boogie program that shows how to encode a Heap
in Strata.Boogie. This will serve as a basis to translate from a Heap
dialect to Strata.Boogie.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: Shilpi Goel <shigoel@gmail.com>