Clarify the interface to the `Map` type (#73)
This is a first step toward clearer interface abstraction so that we can
replace `Map` with a more efficient implementation.

  * Use `def` instead of `abbrev` for `Map`
  * Remove unused `Map` theorems, so we won't need to reprove them
  * Add `ListMap` type
* Use `ListMap` for path conditions, function signatures, and procedure
signatures

There's a little bit of unfortunate duplication, since path conditions
can no longer use the `Maps` type, but it should be possible to resolve
that as we continue to refactor things.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: Andrew Wells <130512013+andrewmwells-amazon@users.noreply.github.com>
Co-authored-by: Shilpi Goel <shigoel@gmail.com>