Python Strata bindings (#86)
This adds a small Python library that can export a Python dialect and
generate Strata from Python ASTs.

This depends on #85 and will be marked as ready for review once that is
merged.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.