Translate `:bvbuiltin` function declarations (#84)
This implements a translation in `BoogieToStrata` that add a body to
each function declared with the `{:bvbuiltin ...}` attribute that uses
the appropriate built-in operator or function in the Boogie.Strata
dialect.

To test the new functionality, this also:
* Changes the output of StrataVerify to make it easier to use in tests.
* Adds parsing and translation of a few new operators to the Boogie
dialect.
* Updates the BoogieToStrata test suite to support expected output
files.
* Adds a Boogie test including bit vector operators with `:bvbuiltin`.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: Shilpi Goel <shigoel@gmail.com>