Add support for multiple triggers (#103)
Now, in the Boogie dialect, it's possible to write quantified formulas
like the following:

    forall x : int, y : int :: { g(x, y) } { f(x) } g(x, y) < f(x);
    forall x : int, y : int :: { g(x, y), f(x) } g(x, y) < f(x);

Each quantifier can have multiple trigger groups, and each group can
have multiple expressions. The indended semantics according to the
SMT-Lib standard is that each group is an alternative pattern that can
cause the quantifier to be instantiated. To instantiate based on a
specific group, the solver should find a ground term that matches each
expression in the group.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: Shilpi Goel <shigoel@gmail.com>