C_Simp -> GOTO (#76)
Connecting CSimp to CBMC (with lots of assumptions). The translation is
from CSimp to JSON. This JSON is then combined with `defaults.json` to
produce a valid input to CBMC.

CBMC can be run with `run_cbmc.sh`. This includes the important step of
adding function preconditions to the `.goto` program.

`test.sh` can be used to check that modifications don't change the
expected JSON (up to code locations).

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.