Python Strata DDM improvements.   (#109)
This migrates the Python dialect to a standalone module so it can be
used independently of generator script and makes improvements to the
core AST types for better usability.


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: Andrew Wells <130512013+andrewmwells-amazon@users.noreply.github.com>
Co-authored-by: Shilpi Goel <shigoel@gmail.com>