Allow solver-specific SMT preludes (#116)
For the moment, this is used only for the Boogie dialect and only when
using Z3. It sets two options that substantially speed up the `Lambda`
and `Quantifiers` Boogie tests.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.