Speed up Lambda type inference and add a `--type-check` option to StrataVerify (#134)
*Description of changes:*

During type inference, we now prune the substitution list for `.app` by
removing the fresh type variable (that stands in for the type of the
`.app` expression) once it's safe to do so. This also includes a proof
that the pruning doesn't violate the well-formedness criteria of the
substitution list.

This step brings around 80% performance improvement vis-a-vis type
inference on certain Boogie benchmarks.

Also, added an option `--type-check` to StrataVerify, which parses and
checks the DDM program and then exits after invoking the semantic
dialect's type check/inference. Contrast this with the pre-existing
`--check` option which does all of the above plus PE and SMT generation
(not SMT solving).

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.