Add UNMODELED to Dyn (#147)
Add `UNMODELED` to Dyn. We'll use this to model language constructs that
don't affect the analysis under consideration. By modeling a source
language construct this way, any analysis downstream of this Dyn program
will treat the construct as a no-op.


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

Co-authored-by: Shilpi Goel <shigoel@gmail.com>