Improve compatibility of Python and Lean DDM implementation (#143)
This is a large PR that makes DDM changes to support source location
information in #112.

* Updates SyntaxCat to use a structure instead of algebraic data type.
This improves efficiency of accessing head.
* ArgDecls is a singleton structure instead of type abbreviation and
coercisions with array no longer allowed. This will allow additional
invariants to be added such as uniqueness of names.
* There is additional checking for undefined names to reduce spurious
errors when module imports cannot be found.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.