Add Laurel, a language for compiling Java to, and in the future also other languages (#152)
## Changes
Introduce a `Laurel` language that serves as a target for compilation
for Java. It contains features that we expect will make it useful for
compilation from other languages as well, such as Python, JavaScript and
TypeScript. However, this first version may contain only all the
features required for compilation from Java. This PR adds a type that
defines the language and an interpreter to define its semantics.

PR contains example programs using a draft syntax that is not defined in
code.

## Testing
No testing of the syntax or semantics yet

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: Aaron Tomb <aarotomb@amazon.com>
Co-authored-by: Shilpi Goel <shigoel@gmail.com>