Add option to remove irrelevant axioms (#194)
Add option `removeIrrelevantAxioms` to detect and remove axioms
determined to be irrelevant. This is set to false by default.

For now, our definition of a "relevant axiom" is quite weak: an axiom
`a` is relevant for a function `f` if `f` occurs in the body of `a`,
including in any trigger expressions. Eventually, we will compute a
transitive closure involving both axioms and functions. However, this
analysis is sufficient to get counterexamples in some cases of interest.

This PR also contains some basic utilities to compute callers/callees
for Boogie functions and procedures.

Note: this PR comes from cherrypicking the Dependency Analysis commits
from [py_to_dyn](https://github.com/strata-org/Strata/pull/148).

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: Shilpi Goel <shilgoel@amazon.com>
Co-authored-by: Shilpi Goel <shigoel@gmail.com>