Replace free vars in SMT DL with UFs (#193)
*Description of changes:*

This PR refactors the handling of SMT variables throughout the codebase,
standardizing the representation of free variables as UFs and
simplifying related logic. The changes remove the `isBound` field from
`TermVar`, update the encoding and lookup of variables, and adjust test
outputs to reflect the changes. The refactor makes the SMT DL more in
line with SMT-LIB and easier to maintain, reducing special-case handling
for variables.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: Shilpi Goel <shigoel@gmail.com>