Add small-step semantics for statements (#200)
The previous semantics are big-step, which is less standard and makes
loops awkward.

To continue to make use of the existing transformation correctness
proofs, we could prove equivalence between the big-step and small-step
semantics (which shouldn't be too hard, but would be some work).

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.