Cleanup Python dialect; add support byte literal constants. (#209)
This adds a DDM Init category Init.ByteArray for byte literals along
with textual and Ion encodings.

It uses this to fix an issue in the Python dialect where byte array
constants were encoded as ellipsis.

This also fixes CI to run again.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: Shilpi Goel <shigoel@gmail.com>