Add `LContext` and operators to typing rules (#203)
Makes `HasType` consistent with behavior of the typechecker by adding
`LContext` and rules for operators. In particular, this PR adds the
condition that constants are well-typed only if their corresoponding
type is known and adds rules for annotated and un-annotated operators
based on their `Factory` instance. It also adds a rule for annotated
free variables. Annotated operators and free variables are somewhat
complex, as they require the annotation to be an instantiation of their
general type.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: Josh Cohen <cohenjo@amazon.com>
Co-authored-by: Shilpi Goel <shigoel@gmail.com>