Add small-step semantics of Lambda (#223)
This pull request adds a small-steps semantics of Lambda! :)

`LExprEvalTests.lean` shows how the small-step semantics evaluates to
the same result, for examples that were already working for concrete
evaluator.

There is an interesting update in the definition if `isCanonicalValue`.

If `e:LExpr` is a series of `.app`, say `e0 e1 .. en`, `e` is a
canonical value if

(1) (_already existed before this PR, added by @joscoh_) `e0` is a
constructor and `e1 .. en` are all canonical values, or
(2) (_newly added_) `e0` is a named function `f` (not abstraction) and
`n` is less than the number of arguments required to run the function
`f`.

The intuition of case (2) is as follows. Let's assume that we would like
to calculate `Int.Add 1 (2+3)`. According to the small step semantics,
we would like to calculate `2+3` to `5`, hence it becomes `Int.Add 1 5`
and eventually 6.
Without (2), this is impossible because the `reduce_2` rule of small
step semantics only fires when `Int.Add 1` is a 'canonical value'.
Therefore, without (2), the semantics stuck and `2+3` can never be
evaluated to `5`.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.