feat: Create generators for well-typed LExprs. (#258)
*Issue #, if available:*

*Description of changes:*

Adds a dependency on Plausible, and creates a number of generators which
ultimately allow generating random `LExpr`s which are well-typed,
labelled with `LMonoTy`s (and random metadata).

Some caveats:

- An instance of `Arbitrary` will need to be supplied for the metadata
types (in our examples we use `Unit` which has a default instance).
- The generators were created using Chamelean
(https://github.com/codyroux/plausible), when possible and cleaned up a
bit by hand. We've removed all the calls to the Chamelean generators, to
avoid taking that dependency, but left them in comments, as
documentation.
- We use a typing relation derived from `LExpr.HasType`, to allow for
easier generation, and remove polymorphism for the time being.
- We show examples of how to use these generators for conformance tests,
but things can be more convenient with additional helper functions.


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.