Add all executables to default targets (#237)
**Description of changes:**
Adding ``strata`` and ``BoogieToGoto`` to default targets list. This
ensures that all binaries are generated when running ``lake build``. My
assumption here is that this was omitted by mistake, but feel free to
discard this PR if that was intentional.




By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

Co-authored-by: Vidas Jocius <vjjocius@users.noreply.github.com>