Change `Imperative.Stmt` to remove `Block` mutual recursion (#216)
Makes some definitions cleaner as they no longer need angle brackets or
`{ss := _}`.

`Block` is now an `abbrev` for `List Stmt`, and we don't need both
`Stmts` and `Block`. This makes the semantics for statements cleaner.

This also includes an induction principle for the new version of `Stmt`,
though it is unused right now.


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: Josh Cohen <cohenjo@amazon.com>
Co-authored-by: Andrew Wells <130512013+andrewmwells-amazon@users.noreply.github.com>
Co-authored-by: Andrew Wells <anmwells@amazon.com>