Fix alpha equivalence for Boogie programs (#265)
Fixes Issue #253

This PR changes alpha-equivalence to build identifier maps between
programs in both directions rather than only one, fixing the problem of
dealing inconsistently with shadowed variable bindings. This is somewhat
more general than what is needed for procedure inlining (e.g. adding the
check that the outputs to procedures are pairwise alpha-equivalent) but
could be useful if alpha equivalence is needed in the future.


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: Josh Cohen <cohenjo@amazon.com>