Strata language definition document (#186)
Adds a document describing the semantics of Strata Core (consisting of
`Lambda` and `Imperative` components). The document is written in Verso
and imports the Strata library to allow docstrings to appear directly in
the text.

Note that Strata Core is not a new dialect, but rather a new name for
the combination of `Lambda` and `Imperative`. It does not yet have a
concrete syntax. Concrete syntax will likely be provided through an
evolution of #224, to assist in the goal of keeping Strata Core as close
to B3 as possible.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: Josh Cohen <cohenjo@amazon.com>
Co-authored-by: Mikaël Mayer <MikaelMayer@users.noreply.github.com>
Co-authored-by: Shilpi Goel <shigoel@gmail.com>
Co-authored-by: Andrew Wells <130512013+andrewmwells-amazon@users.noreply.github.com>