Start porting DDM to use modules (#287)
This migrates many of the DDM Lean source files to use Lean modules.

The Lean module system is a relatively new feature that is no longer
considered experimental. For more information, including the benefits of
the module system, see the [Lean
documentation](https://lean-lang.org/doc/reference/latest/Source-Files-and-Modules/#module-scopes).

I plan to eventually convert all of the DDM source to use modules, but
that can be done incrementally, and the PR is already quite large.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.